Don't use IInitializeManager just to get capabilities
