+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/IWorkspaceSemanticTokensRefreshPublisher.cs
@@ -1,0 +1,9 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT license. See License.txt in the project root for license information.
+
+namespace Microsoft.AspNetCore.Razor.LanguageServer;
+
+internal interface IWorkspaceSemanticTokensRefreshPublisher
+{
+    void EnqueueWorkspaceSemanticTokensRefresh();
+}
